+++ app/src/main/java/org/oxycblt/auxio/music/excluded/ExcludedDatabase.kt
@@ -30,7 +30,9 @@ import org.oxycblt.auxio.util.requireBackgroundThread
  * Database for storing excluded directories. Note that the paths stored here will not work with
  * MediaStore unless you append a "%" at the end. Yes. I know Room exists. But that would needlessly
  * bloat my app and has crippling bugs.
- * @author OxygenCobalt TODO: Migrate this to SharedPreferences?
+ * @author OxygenCobalt
+ *
+ * TODO: Migrate this to SharedPreferences?
  */
 class ExcludedDatabase(context: Context) : SQLiteOpenHelper(context, DB_NAME, null, DB_VERSION) {
     override fun onCreate(db: SQLiteDatabase) {